refactored Hypothesis impls into lib.rs and spruced up docs.
